+++ ddd-core/pom.xml
