+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/domain/AbstractIdentity.java
