+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/event/Event.java
@@ -1,0 +1,7 @@
+package xyz.zhangyi.ddd.eas.core.event;
+
+import java.io.Serializable;
+
+public interface Event extends Serializable {
+    String eventId();
+}
