+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/ApplicationException.java
@@ -1,0 +1,11 @@
+package xyz.zhangyi.ddd.eas.core.exception;
+
+public abstract class ApplicationException extends RuntimeException {
+    public ApplicationException(String message) {
+        super(message);
+    }
+
+    public ApplicationException(String message, Exception ex) {
+        super(message, ex);
+    }
+}
