+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/ApplicationInfrastructureException.java
@@ -1,0 +1,9 @@
+package xyz.zhangyi.ddd.eas.core.exception;
+
+import xyz.zhangyi.ddd.eas.core.exception.ApplicationException;
+
+public class ApplicationInfrastructureException extends ApplicationException {
+    public ApplicationInfrastructureException(String message, Exception ex) {
+        super(message, ex);
+    }
+}
