+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/ApplicationValidationException.java
@@ -1,0 +1,9 @@
+package xyz.zhangyi.ddd.eas.core.exception;
+
+import xyz.zhangyi.ddd.eas.core.exception.ApplicationException;
+
+public class ApplicationValidationException extends ApplicationException {
+    public ApplicationValidationException(String message) {
+        super(message);
+    }
+}
