+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/DomainException.java
@@ -1,0 +1,11 @@
+package xyz.zhangyi.ddd.eas.core.exception;
+
+public class DomainException extends RuntimeException {
+    public DomainException(String message) {
+        super(message);
+    }
+
+    public DomainException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
