+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/gateway/acl/port/EventPublisher.java
@@ -1,0 +1,7 @@
+package xyz.zhangyi.ddd.eas.core.gateway.acl.port;
+
+import xyz.zhangyi.ddd.eas.core.event.Event;
+
+public interface EventPublisher<T extends Event> {
+    void publish(T event);
+}
