--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/application/exception/ApplicationDomainException.java
@@ -1,7 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.application.exception;
-
-public class ApplicationDomainException extends ApplicationException {
-    public ApplicationDomainException(String message, Exception ex) {
-        super(message, ex);
-    }
-}
