--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/application/exception/ApplicationException.java
@@ -1,11 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.application.exception;
-
-public abstract class ApplicationException extends RuntimeException {
-    public ApplicationException(String message) {
-        super(message);
-    }
-
-    public ApplicationException(String message, Exception ex) {
-        super(message, ex);
-    }
-}
