--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/application/exception/ApplicationInfrastructureException.java
@@ -1,7 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.application.exception;
-
-public class ApplicationInfrastructureException extends ApplicationException {
-    public ApplicationInfrastructureException(String message, Exception ex) {
-        super(message, ex);
-    }
-}
