--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/application/exception/ApplicationValidationException.java
@@ -1,7 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.application.exception;
-
-public class ApplicationValidationException extends ApplicationException {
-    public ApplicationValidationException(String message) {
-        super(message);
-    }
-}
