--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/domain/Event.java
@@ -1,7 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.domain;
-
-import java.io.Serializable;
-
-public interface Event extends Serializable {
-    String eventId();
-}
