--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/domain/exception/DomainException.java
@@ -1,11 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.domain.exception;
-
-public class DomainException extends RuntimeException {
-    public DomainException(String message) {
-        super(message);
-    }
-
-    public DomainException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
