--- eas-core/src/main/java/xyz/zhangyi/ddd/eas/core/gateway/acl/port/EventPublisher.java
@@ -1,7 +1,0 @@
-package xyz.zhangyi.ddd.eas.core.gateway.acl.port;
-
-import xyz.zhangyi.ddd.eas.core.domain.Event;
-
-public interface EventPublisher<T extends Event> {
-    void publish(T event);
-}
